Module fp::matrix

Modules

affine 🔒

m4ri 🔒

matrix_inner 🔒

quasi_inverse 🔒

subquotient 🔒

subspace 🔒

Structs

AffineSubspace

AugmentedMatrix

This models an augmented matrix.

Matrix

A matrix! In particular, a matrix with values in F_p. The way we store matrices means it is easier to perform row operations than column operations, and the way we use matrices means we want our matrices to act on the right. Hence we think of vectors as row vectors.

MatrixSliceMut

QuasiInverse

Given a matrix M, a quasi-inverse Q is a map from the co-domain to the domain such that xQM = x for all x in the image (recall our matrices act on the right).

Subquotient

Subspace

A subspace of a vector space.