Struct fp::matrix::quasi_inverse::QuasiInverse

pub struct QuasiInverse {
    image: Option<Vec<isize>>,
    preimage: Matrix,
}

Given a matrix M, a quasi-inverse Q is a map from the co-domain to the domain such that xQM = x for all x in the image (recall our matrices act on the right).

Fields

image - The image of the original matrix. If the image is omitted, it is assumed to be everything (with the standard basis).

preimage - The actual quasi-inverse, where the basis of the image is that given by image.

Fields
image: Option<Vec<isize>>preimage: Matrix
Implementations

impl QuasiInverse

pub fn new(image: Option<Vec<isize>>, preimage: Matrix) -> Self

pub fn image_dimension(&self) -> usize

pub fn source_dimension(&self) -> usize

pub fn target_dimension(&self) -> usize

pub fn to_bytes(&self, buffer: &mut impl Write) -> Result<()>

pub fn from_bytes(p: ValidPrime, data: &mut impl Read) -> Result<Self>

pub fn stream_quasi_inverse<T, S>( p: ValidPrime, data: &mut impl Read, results: &mut [T], inputs: &[S] ) -> Result<()>
where for<'a> &'a mut T: Into<SliceMut<'a>>, for<'a> &'a S: Into<Slice<'a>>,

Given a data file containing a quasi-inverse, apply it to all the vectors in input and write the results to the corresponding vectors in results. This reads in the quasi-inverse row by row to minimize memory usage.

pub fn preimage(&self) -> &Matrix

pub fn pivots(&self) -> Option<&[isize]>

pub fn prime(&self) -> ValidPrime

pub fn apply(&self, target: SliceMut<'_>, coeff: u32, input: Slice<'_>)

Apply the quasi-inverse to an input vector and add a constant multiple of the result to an output vector

Arguments

target - The output vector

coeff - The constant multiple above

input - The input vector, expressed in the basis of the ambient space