Struct fp::matrix::AugmentedMatrix3

pub struct AugmentedMatrix3 {
    pub end: [usize; 3],
    pub start: [usize; 3],
    pub inner: Matrix,
}

This models an augmented matrix.

In an ideal world, this will have no public fields. The inner matrix can be accessed via deref, and there are functions that expose end and start. However, in the real world, the borrow checker exists, and there are cases where directly accessing these fields is what it takes to let you pass the borrow checker.

In particular, if m is an augmented matrix and f is a function that takes in &mut Matrix, trying to run m.f(m.start[0]) produces an error because it is not clear if we first do the deref_mut then retrieve start[0]. (since deref_mut takes in a mutable borrow, it could in theory modify m non-trivially)

Fields
end: [usize; 3]start: [usize; 3]inner: Matrix
Methods

impl AugmentedMatrix3

pub fn new(p: ValidPrime, rows: usize, columns: &[usize]) -> Self

pub fn segment(&mut self, start: usize, end: usize) -> MatrixSlice

pub fn row_segment(
&mut self,
i: usize,
start: usize,
end: usize
) -> FpVectorSlice

pub fn into_matrix(self) -> Matrix

impl AugmentedMatrix3

pub fn compute_quasi_inverses(
&mut self,
pivots: &[isize]
) -> (QuasiInverse, QuasiInverse)