Module fp::prime

Structs

BinomialIterator
Iterates through all numbers with the same number of bits. It may panic or return nonsense after all valid values are returned.

BitflagIterator

ValidPrime

Constants

BINOMIAL4_TABLE

BINOMIAL4_TABLE_SIZE

INVERSE_TABLE

MAX_MULTINOMIAL_LEN

MAX_PRIME

NOT_A_PRIME

NUM_PRIMES

PRIMES

PRIME_TO_INDEX_MAP

Statics

BINOMIAL_TABLE

Traits

Binomial
A number satisfying the Binomial trait supports computing various binomial coefficients. This is implemented using a macro, since the implementation for all types is syntactically the same.

Functions

direct_binomial⚠
This uses a lookup table for n choose k when n and k are both less than p. Lucas’s theorem reduces general binomial coefficients to this case.

integer_power
Computes b^e.

inverse

is_valid_prime

log2
Compute the base 2 log of a number, rounded down to the nearest integer.

logp

minus_one_to_the_n

power_mod
Compute b^e mod p.