fp::prime
Module binomial

Macros

impl_binomial 🔒

Traits

Binomial

A number satisfying the Binomial trait supports computing various binomial coefficients. This is implemented using a macro, since the implementation for all types is syntactically the same.

Functions

direct_binomial 🔒 ⚠

This uses a lookup table for n choose k when n and k are both less than p. Lucas’s theorem reduces general binomial coefficients to this case.