Module fp::vector

An FpVector is a vector with entries in Fp. We use this instead of Vec<u32> because we can pack a lot of entries into a single u64, especially for p small. This not only saves memory, but also leads to faster addition, for example (e.g. a single ^ can add 64 elements of F2 at the same time).

The organization of this file is a bit funny. There are in fact 4 different implementations of FpVector — the cases p = 2, 3, 5 are handled separately with some extra documentation. Hence there are structs FpVector2, FpVector3, FpVector5 and FpVectorGeneric. FpVector itself is an enum that can be either of these. All of these implement the trait FpVectorT, which is where most functions lie. The implementations for FpVector of course just calls the implementations of the specific instances, and this is automated via enum_dispatch.

To understand the methods of FpVector, one should mostly look at the documentation for FpVectorT. However, the static functions for FpVector are implemented in FpVector itself, and hence is documented there as well. The documentation of FpVector2, FpVector3, FpVector5, FpVectorGeneric are basically useless (and empty).

In practice, one only ever needs to work with the enum FpVector and the associated functions. However, the way this structured means one always has to import both FpVector and FpVectorT, since you cannot use the functions of a trait unless you have imported the trait.

Structs

AddShiftLeftData

AddShiftNoneData

AddShiftRightData

FpVector2

FpVector2IteratorNonzero

FpVector3

FpVector3IteratorNonzero

FpVector5

FpVector5IteratorNonzero

FpVectorGeneric

FpVectorGenericIteratorNonzero

FpVectorIterator

FpVectorMask
An FpVectorMask encodes a subset of the basis elements of an Fp vector space. This is used to project onto the subspace spanned by the selected basis elements.

FpVectorSlice

LimbBitIndexPair

VectorContainer

VectorDiffEntry

Enums

FpVector

FpVectorIteratorNonzero

Constants

MAX_DIMENSION

Statics

BITMASKS

BIT_LENGHTS

ENTRIES_PER_64_BITS

LIMB_BIT_INDEX_ONCE_TABLE

LIMB_BIT_INDEX_TABLE
This table tells us which limb and which bitfield of that limb to look for a given index of the vector in.

Traits

FpVectorT

Functions

bit_length

bitmask

entries_per_64_bits

initialize_limb_bit_index_table

limb_bit_index_pair