fp::vector::fp_wrapper
Module macros_generic

Macros

dispatch_struct 🔒

Define a struct that wraps some field-dependent type in an enum. Since Fp<P2> through Fp<P7> are all ZSTs, working with them under the hood is in theory much faster than working with an arbitrary prime field, let alone finite field.

dispatch_vector 🔒

dispatch_vector_inner 🔒

impl_from 🔒

impl_from_inner 🔒

impl_try_into 🔒

impl_try_into_inner 🔒

use_primes 🔒

Use all primes. It isn’t possible to have this import when odd-primes is disabled because primes other than P2 and ValidPrime (which happen to be the same type) do not exist.