Trait fp::vector::FpVectorT

pub trait FpVectorT {
    fn all_leq_limb(&self, limb_a: u64, limb_b: u64) -> bool;

    fn is_reduced_limb(&self, limb: u64) -> bool;

    fn reduce_limb(&self, limb: u64) -> u64;

    fn reduce_quotient_limb(&self, limb: u64) -> (u64, u64);

    fn vector_container(&self) -> &VectorContainer;

    fn vector_container_mut(&mut self) -> &mut VectorContainer;

    fn prime(&self) -> ValidPrime;

    fn add_limb(&self, limb_a: u64, limb_b: u64, coeff: u32) -> u64 { ... }

    fn all_leq(&self, other: &FpVector) -> bool { ... }

    fn all_leq_shift_none(&self, other: &FpVector) -> bool { ... }

    fn all_leq_shift_left(&self, _other: &FpVector) -> bool { ... }

    fn all_leq_shift_right(&self, _other: &FpVector) -> bool { ... }

    fn add(&mut self, other: &FpVector, c: u32) { ... }

    fn add_shift_none_pure(&mut self, other: &FpVector, c: u32) { ... }

    fn add_shift_none(&mut self, other: &FpVector, c: u32) { ... }

    fn add_shift_left(&mut self, other: &FpVector, c: u32) { ... }

    fn add_shift_right(&mut self, other: &FpVector, c: u32) { ... }

    fn truncate_limb(&self, sum: u64) -> Option<u64> { ... }

    fn add_truncate(&mut self, other: &FpVector, c: u32) -> Option<()> { ... }

    fn add_truncate_shift_none(
        &mut self, 
        other: &FpVector, 
        c: u32
    ) -> Option<()> { ... }

    fn add_truncate_shift_left(
        &mut self, 
        other: &FpVector, 
        c: u32
    ) -> Option<()> { ... }

    fn add_truncate_shift_right(
        &mut self, 
        other: &FpVector, 
        c: u32
    ) -> Option<()> { ... }

    fn reduce_limbs(&mut self, start_limb: usize, end_limb: usize) { ... }

    fn dimension(&self) -> usize { ... }

    fn offset(&self) -> usize { ... }

    fn min_index(&self) -> usize { ... }

    fn slice(&self) -> (usize, usize) { ... }

    fn is_set_slice_valid(&self, slice_start: usize, slice_end: usize) -> bool { ... }

    fn set_slice(&mut self, slice_start: usize, slice_end: usize) { ... }

    fn restore_slice(&mut self, slice: (usize, usize)) { ... }

    fn clear_slice(&mut self) { ... }

    fn into_slice(&mut self) { ... }

    fn min_limb(&self) -> usize { ... }

    fn max_limb(&self) -> usize { ... }

    fn limbs(&self) -> &Vec<u64> { ... }

    fn limbs_mut(&mut self) -> &mut Vec<u64> { ... }

    fn take_limbs(&mut self) -> Vec<u64> { ... }

    fn put_limbs(&mut self, limbs: Vec<u64>) { ... }

    fn limb_mask(&self, limb_idx: usize) -> u64 { ... }

    fn set_to_zero_pure(&mut self) { ... }

    fn set_to_zero(&mut self) { ... }

    fn shift_assign(&mut self, other: &FpVector) { ... }

    fn assign(&mut self, other: &FpVector) { ... }

    fn is_zero_pure(&self) -> bool { ... }

    fn is_zero(&self) -> bool { ... }

    fn entry(&self, index: usize) -> u32 { ... }

    fn set_entry(&mut self, index: usize, value: u32) { ... }

    fn add_basis_element(&mut self, index: usize, value: u32) { ... }

    fn unpack(&self, target: &mut [u32]) { ... }

    fn to_vector(&self) -> Vec<u32> { ... }

    fn pack(&mut self, source: &[u32]) { ... }

    fn add_tensor(
        &mut self, 
        offset: usize, 
        coeff: u32, 
        left: &FpVector, 
        right: &FpVector
    ) { ... }

    fn scale(&mut self, c: u32) { ... }
}

Required methods

fn all_leq_limb(&self, limb_a: u64, limb_b: u64) -> bool

fn is_reduced_limb(&self, limb: u64) -> bool

fn reduce_limb(&self, limb: u64) -> u64

fn reduce_quotient_limb(&self, limb: u64) -> (u64, u64)

fn vector_container(&self) -> &VectorContainer

fn vector_container_mut(&mut self) -> &mut VectorContainer

fn prime(&self) -> ValidPrime

Loading content...
Provided methods

fn add_limb(&self, limb_a: u64, limb_b: u64, coeff: u32) -> u64

fn all_leq(&self, other: &FpVector) -> bool

fn all_leq_shift_none(&self, other: &FpVector) -> bool

fn all_leq_shift_left(&self, _other: &FpVector) -> bool

fn all_leq_shift_right(&self, _other: &FpVector) -> bool

fn add(&mut self, other: &FpVector, c: u32)

fn add_shift_none_pure(&mut self, other: &FpVector, c: u32)

Ignores any slice and any edge conditions. For use in row_reduce. Included because profiling showed that handling edge conditions was taking up 2%

fn add_shift_none(&mut self, other: &FpVector, c: u32)

Adds c * other to self. other must have the same length, offset, and prime as self, and c must be between 0 and p - 1.

fn add_shift_left(&mut self, other: &FpVector, c: u32)

fn add_shift_right(&mut self, other: &FpVector, c: u32)

fn truncate_limb(&self, sum: u64) -> Option<u64>

fn add_truncate(&mut self, other: &FpVector, c: u32) -> Option<()>

fn add_truncate_shift_none(&mut self, other: &FpVector, c: u32) -> Option<()>

Adds c * other to self. other must have the same length, offset, and prime as self, and c must be between 0 and p - 1. If any of the fields exceeds p after doing this, return “false” and quit as soon as this condition is detected. In this case, “self” will contain undefined nonsense. Otherwise return “true” and “self” will contain the sum. You get these “_truncate” variants from the normal variants by: every time “self.add_limb()” shows up in the original variant, replace it with “self.add_limb_truncate()?”. Also have to add some extra Some(())’s. Adds c * other to self. other must have the same length, offset, and prime as self, and c must be between 0 and p - 1.

fn add_truncate_shift_left(&mut self, other: &FpVector, c: u32) -> Option<()>

fn add_truncate_shift_right(&mut self, other: &FpVector, c: u32) -> Option<()>

fn reduce_limbs(&mut self, start_limb: usize, end_limb: usize)

fn dimension(&self) -> usize

fn offset(&self) -> usize

fn min_index(&self) -> usize

fn slice(&self) -> (usize, usize)

fn is_set_slice_valid(&self, slice_start: usize, slice_end: usize) -> bool

fn set_slice(&mut self, slice_start: usize, slice_end: usize)

fn restore_slice(&mut self, slice: (usize, usize))

fn clear_slice(&mut self)

fn into_slice(&mut self)

Drops every element in the fp_vector that is not in the current slice.

fn min_limb(&self) -> usize

fn max_limb(&self) -> usize

fn limbs(&self) -> &Vec<u64>

fn limbs_mut(&mut self) -> &mut Vec<u64>

fn take_limbs(&mut self) -> Vec<u64>

fn put_limbs(&mut self, limbs: Vec<u64>)

fn limb_mask(&self, limb_idx: usize) -> u64

fn set_to_zero_pure(&mut self)

fn set_to_zero(&mut self)

fn shift_assign(&mut self, other: &FpVector)

fn assign(&mut self, other: &FpVector)

fn is_zero_pure(&self) -> bool

fn is_zero(&self) -> bool

fn entry(&self, index: usize) -> u32

fn set_entry(&mut self, index: usize, value: u32)

fn add_basis_element(&mut self, index: usize, value: u32)

fn unpack(&self, target: &mut [u32])

Unpacks an FpVector onto an array slice. note that the array slice has to be long enough to hold all the elements in the FpVector.

fn to_vector(&self) -> Vec<u32>

fn pack(&mut self, source: &[u32])

fn add_tensor(
&mut self,
offset: usize,
coeff: u32,
left: &FpVector,
right: &FpVector
)

coeff need not be reduced mod p. Adds v otimes w to self.

fn scale(&mut self, c: u32)

Loading content...