Module fp::vector_inner::limb

Functions

add

is_reduced

number

pack
Given an interator of u32’s, pack all of them into a single limb in order. It is assumed that

range

reduce
Contbuted by Robert Burklund

sign_rule

truncate
Returns: either Some(sum) if no carries happen in the limb or None if some carry does happen.

unpack
Given a limb, return the first dim entries. It is assumed that dim is not greater than the number of entries in a limb.