Struct fp::vector_inner::SliceMutP

pub struct SliceMutP<'a, const P: u32> {
    limbs: &'a mut [u64],
    start: usize,
    end: usize,
}

A SliceMutP is a mutable slice of an FpVectorP. This mutably borrows the vector. Since it is a mutable borrow, it cannot implement Copy. However, it has a SliceMutP::copy function that imitates the reborrowing, that mutably borrows SliceMutP and returns a SliceMutP with a shorter lifetime.

Fields
limbs: &'a mut [u64]start: usizeend: usize
Implementations

impl<'a, const P: u32> SliceMutP<'a, P>

pub fn slice_mut(&mut self, start: usize, end: usize) -> SliceMutP<'_, P>

pub fn as_slice(&self) -> SliceP<'_, P>

pub fn copy(&mut self) -> SliceMutP<'_, P>

Generates a version of itself with a shorter lifetime

impl<'a, const P: u32> SliceMutP<'a, P>

pub fn prime(&self) -> ValidPrime

pub fn add_basis_element(&mut self, index: usize, value: u32)

pub fn set_entry(&mut self, index: usize, value: u32)

fn reduce_limbs(&mut self)

pub fn scale(&mut self, c: u32)

pub fn set_to_zero(&mut self)

pub fn add(&mut self, other: SliceP<'_, P>, c: u32)

pub fn add_tensor( &mut self, offset: usize, coeff: u32, left: SliceP<'_, P>, right: SliceP<'_, P> )

coeff need not be reduced mod p. Adds v otimes w to self.

pub fn assign(&mut self, other: SliceP<'_, P>)

TODO: improve efficiency

pub fn add_shift_none(&mut self, other: SliceP<'_, P>, c: u32)

Adds c * other to self. other must have the same length, offset, and prime as self, and c must be between 0 and p - 1.

fn add_shift_left(&mut self, other: SliceP<'_, P>, c: u32)

fn add_shift_right(&mut self, other: SliceP<'_, P>, c: u32)

pub fn add_masked(&mut self, other: SliceP<'_, P>, c: u32, mask: &[usize])

Given a mask v, add the v[i]th entry of other to the ith entry of self.

pub fn add_unmasked(&mut self, other: SliceP<'_, P>, c: u32, mask: &[usize])

Given a mask v, add the ith entry of other to the v[i]th entry of self.