Struct fp::vector_inner::SliceP

pub struct SliceP<'a, const P: u32> {
    limbs: &'a [u64],
    start: usize,
    end: usize,
}

A SliceP is a slice of an FpVectorP. This immutably borrows the vector and implements Copy

Fields
limbs: &'a [u64]start: usizeend: usize
Implementations

impl<'a, const P: u32> SliceP<'a, P>

pub fn slice(self, start: usize, end: usize) -> SliceP<'a, P>

pub fn to_owned(self) -> FpVectorP<P>

Converts a slice to an owned FpVectorP. This is vastly more efficient if the start of the vector is aligned.

impl<'a, const P: u32> SliceP<'a, P>

pub fn prime(&self) -> ValidPrime

pub fn len(&self) -> usize

pub const fn is_empty(&self) -> bool

pub fn entry(&self, index: usize) -> u32

pub fn iter(self) -> FpVectorIterator<'a>

TODO: implement prime 2 version

pub fn iter_nonzero(self) -> FpVectorNonZeroIteratorP<'a, P>

pub fn is_zero(&self) -> bool

impl<'a, const P: u32> SliceP<'a, P>

fn offset(&self) -> usize

fn limb_range(&self) -> Range<usize>

fn limb_range_inner(&self) -> Range<usize>

This function underflows if self.end == 0, which happens if and only if we are taking a slice of width 0 at the start of an FpVector. This should be a very rare edge case. Dealing with the underflow properly would probably require using saturating_sub or something of that nature, and that has a nontrivial (10%) performance hit.

fn min_limb_mask(&self) -> u64

fn max_limb_mask(&self) -> u64

fn limb_masks(&self) -> (u64, u64)