Crate fp

Modules

matrix

prime

vector
This module is provides wrappers around the contents of crate::vector_inner. The main purpose is to put FpVectorP for different p into a single enum. It does the same for the various slice structs.

vector_2
This module replaces vector when odd-primes is disabled. Instead of producing enum wrappers, it simply rexports FooP<2> as Foo.

vector_inner

Macros

const_for

Traits

TryInto
This is a version of std::convert::TryInto that deals with mutable references instead. We cannot condition on TryInto for mutable references due to lifetime issues. This is used for the add_carry implementation