Crate fp

Modules

constants 🔒

field

limb 🔒

matrix

prime

simd 🔒

vector

Macros

const_for

Constants

MAX_MULTINOMIAL_LEN

NUM_PRIMES

The number of primes that will be prepared at compile-time. This includes precomputing inverses, binomial coefficients, etc.

ODD_PRIMES

PRIMES

An array containing the first NUM_PRIMES prime numbers.

PRIME_TO_INDEX_MAP

For any integer i less than or equal to MAX_PRIME, PRIME_TO_INDEX_MAP[i] is the index of i in PRIMES if i is prime; otherwise, it is NOT_A_PRIME.

Traits

MaybeArbitrary 🔒