Crate fp

Modules

constants 🔒

limb 🔒

matrix

prime

simd 🔒

vector

This module is provides wrappers around the contents of crate::vector_inner. The main purpose is to put FpVectorP for different p into a single enum. It does the same for the various slice structs.

vector_2

This module replaces vector when odd-primes is disabled. Instead of producing enum wrappers, it simply rexports FooP<2> as Foo.

vector_inner

Macros

const_for

Constants

MAX_MULTINOMIAL_LEN

NUM_PRIMES

The number of primes that are supported.

ODD_PRIMES

PRIMES

An array containing the first NUM_PRIMES prime numbers.

PRIME_TO_INDEX_MAP

For any integer i less than or equal to MAX_PRIME, PRIME_TO_INDEX_MAP[i] is the index of i in PRIMES if i is prime; otherwise, it is NOT_A_PRIME.