fp
Module constants

Macros

populate_binomial_table 🔒

Constants

BINOMIAL4_TABLE 🔒

BINOMIAL4_TABLE_SIZE 🔒

BITS_PER_LIMB

The number of bits each Limb occupies.

BYTES_PER_LIMB

The number of bytes each Limb occupies.

MAX_MULTINOMIAL_LEN

MAX_PRIME

The NUM_PRIMESth prime number.

NOT_A_PRIME

A sentinel value. PRIME_TO_INDEX_MAP[i] == NOT_A_PRIME if and only if i is not a prime number.

NUM_PRIMES

The number of primes that will be prepared at compile-time. This includes precomputing inverses, binomial coefficients, etc.

PRIMES

An array containing the first NUM_PRIMES prime numbers.

PRIME_TO_INDEX_MAP

For any integer i less than or equal to MAX_PRIME, PRIME_TO_INDEX_MAP[i] is the index of i in PRIMES if i is prime; otherwise, it is NOT_A_PRIME.

Statics

BINOMIAL_TABLE 🔒

INVERSE_TABLE 🔒

Type Aliases

Limb 🔒