Module fp::constants

Macros

populate_binomial_table 🔒

Constants

BINOMIAL4_TABLE 🔒

BINOMIAL4_TABLE_SIZE 🔒

BITS_PER_LIMB

The number of bits each Limb occupies.

BYTES_PER_LIMB

The number of bytes each Limb occupies.

INVERSE_TABLE 🔒

MAX_MULTINOMIAL_LEN

MAX_PRIME

The MAX_PRIMEth prime number. Constructing a ValidPrime using any number larger than this value will cause a panic.

NOT_A_PRIME

A sentinel value. PRIME_TO_INDEX_MAP[i] == NOT_A_PRIME if and only if i is not a prime number.

NUM_PRIMES

The number of primes that are supported.

PRIMES

An array containing the first NUM_PRIMES prime numbers.

PRIME_TO_INDEX_MAP

For any integer i less than or equal to MAX_PRIME, PRIME_TO_INDEX_MAP[i] is the index of i in PRIMES if i is prime; otherwise, it is NOT_A_PRIME.

Statics

BINOMIAL_TABLE 🔒

Type Definitions

Limb 🔒