fp::field

Module smallfq

Structs

SmallFq

A field of order q = p^d, where q < 2^16 and d > 1. Fields of that size are small enough that we can cache their Zech logarithms.

SmallFqElement

A field element, stored as the exponent of a distinguished generator of the group of units. None if the element is zero.

Statics

SMALL_CONWAY_POLYS 🔒

ZECH_LOGS 🔒

A table of lazily initialized Zech logarithms.

Functions

make_zech_log_table 🔒

mul_by_a 🔒

zech_logs 🔒

Return the Zech logarithm table for the given field. If it does not exist yet, initialize it. The initialization might be fairly expensive (several ms).

Type Aliases

Polynomial 🔒

ZechTable 🔒