Struct fp::matrix::matrix_inner::AugmentedMatrix

pub struct AugmentedMatrix<const N: usize> {
    pub end: [usize; N],
    pub start: [usize; N],
    pub inner: Matrix,
}

This models an augmented matrix.

In an ideal world, this will have no public fields. The inner matrix can be accessed via deref, and there are functions that expose end and start. However, in the real world, the borrow checker exists, and there are cases where directly accessing these fields is what it takes to let you pass the borrow checker.

In particular, if m is an augmented matrix and f is a function that takes in &mut Matrix, trying to run m.f(m.start[0]) produces an error because it is not clear if we first do the deref_mut then retrieve start[0]. (since deref_mut takes in a mutable borrow, it could in theory modify m non-trivially)

Fields
end: [usize; N]start: [usize; N]inner: Matrix
Implementations

impl<const N: usize> AugmentedMatrix<N>

pub fn new(p: ValidPrime, rows: usize, columns: [usize; N]) -> Self

pub fn new_with_capacity( p: ValidPrime, rows: usize, columns: &[usize], row_capacity: usize, extra_column_capacity: usize, ) -> Self

pub fn segment(&mut self, start: usize, end: usize) -> MatrixSliceMut<'_>

pub fn row_segment_mut( &mut self, i: usize, start: usize, end: usize, ) -> SliceMut<'_>

pub fn row_segment(&self, i: usize, start: usize, end: usize) -> Slice<'_>

pub fn into_matrix(self) -> Matrix

pub fn into_tail_segment( self, row_start: usize, row_end: usize, segment_start: usize, ) -> Matrix

pub fn compute_kernel(&self) -> Subspace

pub fn extend_column_dimension(&mut self, columns: usize)

impl AugmentedMatrix<2>

pub fn compute_image(&self) -> Subspace

pub fn compute_quasi_inverse(&self) -> QuasiInverse

impl AugmentedMatrix<3>

pub fn drop_first(self) -> AugmentedMatrix<2>

pub fn compute_quasi_inverses(self) -> (QuasiInverse, QuasiInverse)

This function computes quasi-inverses for matrices A, B given a reduced row echelon form of [A|0|B|0|I] such that A is surjective. Moreover, if Q is the quasi-inverse of A, it is guaranteed that the image of QB and B|_{ker A} are disjoint.

This takes ownership of the matrix since it heavily modifies the matrix. This is not strictly necessary but is fine in most applications.