Struct fp::matrix::Subquotient

pub struct Subquotient {
    gens: Subspace,
    quotient: Subspace,
    dimension: usize,
}

Fields
gens: Subspacequotient: Subspacedimension: usize
Implementations

impl Subquotient

pub fn new(p: ValidPrime, dim: usize) -> Self

Create a new subquotient of an ambient space of dimension dim. This defaults to the zero subspace.

pub fn new_full(p: ValidPrime, dim: usize) -> Self

Create a new subquotient of an ambient space of dimension dim, where the subspace is the full subspace and quotient is trivial.

pub fn reduce(&self, elt: SliceMut<'_>) -> Vec<u32>

Given a vector elt, project elt to the complement and express as a linear combination of the basis. The result is returned as a list of coefficients. If elt is nonzero afterwards, this means the vector was not in the subspace to begin with.

pub fn reduce_by_quotient(&self, elt: SliceMut<'_>)

Project the vector onto the complement of the quotient part of the subquotient.

pub fn set_to_full(&mut self)

Set the subquotient to be the full ambient space quotiented by zero

pub fn zeros(&self) -> &Subspace

pub fn gens(&self) -> impl Iterator<Item = &FpVector>

pub fn subspace_dimension(&self) -> usize

The dimension of the subspace part of the subquotient.

pub fn subspace_gens(&self) -> impl Iterator<Item = &FpVector>

The generators of the subspace part of the subquotient.

pub fn quotient(&mut self, elt: Slice<'_>)

pub fn dimension(&self) -> usize

pub fn ambient_dimension(&self) -> usize

pub fn prime(&self) -> ValidPrime

pub fn is_empty(&self) -> bool

pub fn clear_gens(&mut self)

pub fn add_gen(&mut self, gen: Slice<'_>)

pub fn reduce_matrix(
matrix: &Matrix,
source: &Self,
target: &Self
) -> Vec<Vec<u32>>

pub fn subquotient(space: &Subspace, subspace: &Subspace) -> Vec<usize>

Given a chain of subspaces subspace < space < k^ambient_dimension, compute the subquotient space/subspace. The answer is expressed as a list of basis vectors of space whose image in space/subspace forms a basis, and a basis vector of space is described by its index in the list of basis vectors of space (not the ambient space).

Arguments

space - If this is None, it is the whole space k^ambient_dimension

subspace - If this is None, it is empty