Struct fp::matrix::Subspace

#[repr(transparent)]
pub struct Subspace {
    pub matrix: Matrix,
}

A subspace of a vector space.

Fields

matrix - A matrix in reduced row echelon, whose number of columns is the dimension of the ambient space and each row is a basis vector of the subspace.

pivots - If the column is a pivot column, the entry is the row the pivot corresponds to. If the column is not a pivot column, this is some negative number — not necessarily -1!

Fields
matrix: Matrix
Implementations

impl Subspace

pub fn new(p: ValidPrime, rows: usize, columns: usize) -> Self

pub fn empty_space(p: ValidPrime, dim: usize) -> Self

pub fn entire_space(p: ValidPrime, dim: usize) -> Self

pub fn add_vector(&mut self, row: Slice<'_>) -> usize

This adds a vector to the subspace. This function assumes that the last row of the matrix is zero, i.e. the dimension of the current subspace is strictly less than the number of rows. This can be achieved by setting the number of rows to be the dimension plus one when creating the subspace.

Returns

The new dimension of the subspace

pub fn add_vectors(&mut self, rows: impl Iterator<Item = FpVector>)

pub fn add_basis_elements(&mut self, rows: impl Iterator<Item = usize>)

pub fn reduce(&self, vector: SliceMut<'_>)

Projects a vector to a complement of the subspace. The complement is the set of vectors that have a 0 in every column where there is a pivot in matrix

pub fn row_reduce(&mut self) -> usize

pub fn contains(&self, vector: Slice<'_>) -> bool

pub fn dimension(&self) -> usize

pub fn ambient_dimension(&self) -> usize

pub fn basis(&self) -> &[FpVector]

Returns a basis of the subspace.

pub fn set_to_zero(&mut self)

Sets the subspace to be the zero subspace.

pub fn set_to_entire(&mut self)

Sets the subspace to be the entire subspace.