Struct fp::matrix::subspace::Subspace

#[repr(transparent)]
pub struct Subspace {
    pub matrix: Matrix,
}

A subspace of a vector space.

In general, a method is defined on the Subspace if it is a meaningful property of the subspace itself. Otherwise, users are expected to access the matrix object directly. When the user directly modifies the matrix, they are expected to ensure the matrix is row reduced after the operations conclude.

Fields

matrix - A matrix in reduced row echelon, whose number of columns is the dimension of the ambient space and each row is a basis vector of the subspace.

Fields
matrix: Matrix
Implementations

impl Subspace

pub fn new(p: ValidPrime, rows: usize, columns: usize) -> Self

pub fn prime(&self) -> ValidPrime

pub fn pivots(&self) -> &[isize]

pub fn from_bytes(p: ValidPrime, data: &mut impl Read) -> Result<Self>

pub fn to_bytes(&self, buffer: &mut impl Write) -> Result<()>

pub fn empty_space(p: ValidPrime, dim: usize) -> Self

pub fn entire_space(p: ValidPrime, dim: usize) -> Self

pub fn add_vector(&mut self, row: Slice<'_>) -> usize

This adds a vector to the subspace. This function assumes that the last row of the matrix is zero, i.e. the dimension of the current subspace is strictly less than the number of rows. This can be achieved by setting the number of rows to be the dimension plus one when creating the subspace.

Returns

The new dimension of the subspace

pub fn add_vectors( &mut self, rows: impl for<'a> FnMut(SliceMut<'a>) -> Option<()> )

This adds some rows to the subspace

Arguments

rows: A function that writes the row to be added to the given SliceMut. This returns None if it runs out of rows, Some(()) otherwise.

pub fn add_basis_elements(&mut self, rows: impl Iterator<Item = usize>)

pub fn reduce(&self, vector: SliceMut<'_>)

Projects a vector to a complement of the subspace. The complement is the set of vectors that have a 0 in every column where there is a pivot in matrix

pub fn contains(&self, vector: Slice<'_>) -> bool

pub fn dimension(&self) -> usize

pub fn is_empty(&self) -> bool

Whether the subspace is empty. This assumes the subspace is row reduced.

pub fn ambient_dimension(&self) -> usize

pub fn basis(&self) -> &[FpVector]

Returns a basis of the subspace.

pub fn set_to_zero(&mut self)

Sets the subspace to be the zero subspace.

pub fn set_to_entire(&mut self)

Sets the subspace to be the entire subspace.

pub fn iter(&self) -> impl Iterator<Item = Slice<'_>>