Struct fp::matrix::subspace::Subspace

#[repr(transparent)]
pub struct Subspace {
    matrix: Matrix,
}

A subspace of a vector space.

In general, a method is defined on the Subspace if it is a meaningful property of the subspace itself. Otherwise, users can dereference the subspace to gain read-only access to the underlying matrix object.

Fields

matrix - A matrix in reduced row echelon, whose number of columns is the dimension of the ambient space and each row is a basis vector of the subspace.

Fields
matrix: Matrix
Implementations

impl Subspace

pub fn new(p: ValidPrime, dim: usize) -> Self

pub fn from_matrix(matrix: Matrix) -> Self

Create a new subspace from a matrix. The matrix does not have to be in row echelon form.

pub fn update_then_row_reduce<T, F: FnOnce(&mut Matrix) -> T>( &mut self, f: F, ) -> T

Run a closure on the matrix and then ensure it is row-reduced.

pub fn from_bytes(p: ValidPrime, data: &mut impl Read) -> Result<Self>

pub fn to_bytes(&self, buffer: &mut impl Write) -> Result<()>

pub fn entire_space(p: ValidPrime, dim: usize) -> Self

pub fn add_vector(&mut self, row: FpSlice<'_>) -> usize

This adds a vector to the subspace. This function assumes that the last row of the matrix is zero, i.e. the dimension of the current subspace is strictly less than the number of rows. This can be achieved by setting the number of rows to be the dimension plus one when creating the subspace.

Returns

The new dimension of the subspace

pub fn add_vectors( &mut self, rows: impl for<'a> FnMut(FpSliceMut<'a>) -> Option<()>, )

This adds some rows to the subspace

Arguments

rows: A function that writes the row to be added to the given FpSliceMut. This returns None if it runs out of rows, Some(()) otherwise.

pub fn add_basis_elements(&mut self, rows: impl Iterator<Item = usize>)

pub fn reduce(&self, vector: FpSliceMut<'_>)

Projects a vector to a complement of the subspace. The complement is the set of vectors that have a 0 in every column where there is a pivot in matrix

pub fn contains(&self, vector: FpSlice<'_>) -> bool

pub fn contains_space(&self, other: &Self) -> bool

pub fn dimension(&self) -> usize

pub fn is_empty(&self) -> bool

Whether the subspace is empty. This assumes the subspace is row reduced.

pub fn ambient_dimension(&self) -> usize

pub fn basis(&self) -> &[FpVector]

Returns a basis of the subspace.

pub fn set_to_zero(&mut self)

Sets the subspace to be the zero subspace.

pub fn set_to_entire(&mut self)

Sets the subspace to be the entire subspace.

pub fn iter(&self) -> impl Iterator<Item = FpSlice<'_>>

pub fn iter_all_vectors(&self) -> impl Iterator<Item = FpVector> + '_

Iterate over all vectors in the subspace.

Example

let subspace = Subspace::from_matrix(Matrix::from_vec(
    ValidPrime::new(3),
    &[vec![1, 0, 0], vec![0, 1, 2]],
));
let mut all_vectors = subspace.iter_all_vectors().map(|v| (&v).into());

assert_eq!(all_vectors.next(), Some(vec![0, 0, 0]));
assert_eq!(all_vectors.next(), Some(vec![0, 1, 2]));
assert_eq!(all_vectors.next(), Some(vec![0, 2, 1]));
assert_eq!(all_vectors.next(), Some(vec![1, 0, 0]));
assert_eq!(all_vectors.next(), Some(vec![1, 1, 2]));
assert_eq!(all_vectors.next(), Some(vec![1, 2, 1]));
assert_eq!(all_vectors.next(), Some(vec![2, 0, 0]));
assert_eq!(all_vectors.next(), Some(vec![2, 1, 2]));
assert_eq!(all_vectors.next(), Some(vec![2, 2, 1]));
assert_eq!(all_vectors.next(), None);

pub fn sum(&self, other: &Self) -> Self