Module fp::prime

Macros

impl_binomial 🔒

Structs

BinomialIterator

Iterates through all numbers with the same number of bits. It may panic or return nonsense after all valid values are returned.

BitflagIterator

InvalidPrimeError

ValidPrime

Constants

TWO

Traits

Binomial

A number satisfying the Binomial trait supports computing various binomial coefficients. This is implemented using a macro, since the implementation for all types is syntactically the same.

Functions

direct_binomial 🔒 ⚠

This uses a lookup table for n choose k when n and k are both less than p. Lucas’s theorem reduces general binomial coefficients to this case.

factor_pk

Factor $n$ as $p^k m$. Returns $(k, m)$.

integer_power

Computes b^e.

inverse

is_valid_prime

log2

Compute the base 2 log of a number, rounded down to the nearest integer.

logp

Discrete log base p of n.

minus_one_to_the_n

power_mod

Compute b^e mod p.