Module fp::vector

This module is provides wrappers around the contents of crate::vector_inner. The main purpose is to put FpVectorP for different p into a single enum. It does the same for the various slice structs.

The main magic occurs in the macro dispatch_vector_inner, which we use to provide wrapper functions around the FpVectorP functions.

This module is only used when the odd-primes feature is enabled.

Macros

dispatch_vector 🔒

dispatch_vector_inner 🔒

impl_try_into 🔒

match_p 🔒

Enums

FpVector

FpVectorNonZeroIterator

Slice

SliceMut