fp
Module vector

Re-exports

pub use inner::*;

Modules

arbitrary

fp_wrapper 🔒

This module provides convenience wrappers around the contents of crate::vector::inner in the special case where the field is a prime field. The main purpose is to put FqVector for different fields Fp<P> into a single enum, and to simplify scalars to just u32s instead of rather unwieldy FieldElement<Fp<P>>s. It does the same for the various slice structs.

impl_fqslice 🔒

impl_fqslicemut 🔒

impl_fqvector 🔒

inner

iter 🔒

Enums

FpSlice

FpSliceMut

FpVector

FpVectorIterator

FpVectorNonZeroIterator