fp::vector::inner
Struct FqSlice

pub struct FqSlice<'a, F: Field> {
    pub(super) fq: F,
    pub(super) limbs: &'a [u64],
    pub(super) start: usize,
    pub(super) end: usize,
}

A slice of an FqVector.

This immutably borrows the vector and implements Copy.

Fields
fq: Flimbs: &'a [u64]start: usizeend: usize
Implementations

impl<F: Field> FqSlice<'_, F>

pub(super) fn entry_helper(&self, index: usize) -> F::ElementContainer

impl<'a, F: Field> FqSlice<'a, F>

pub fn prime(&self) -> ValidPrime

pub fn len(&self) -> usize

pub const fn is_empty(&self) -> bool

pub fn entry(&self, index: usize) -> FieldElement<F>

pub fn iter(self) -> FqVectorIterator<'a, F>

TODO: implement prime 2 version

pub fn iter_nonzero(self) -> FqVectorNonZeroIterator<'a, F>

pub fn is_zero(&self) -> bool

pub fn slice(self, start: usize, end: usize) -> Self

pub fn to_owned(self) -> FqVector<F>

Converts a slice to an owned FqVector. This is vastly more efficient if the start of the vector is aligned.

impl<F: Field> FqSlice<'_, F>

pub(super) fn offset(&self) -> usize

pub(super) fn limb_range(&self) -> Range<usize>

pub(super) fn limb_range_inner(&self) -> Range<usize>

This function underflows if self.end == 0, which happens if and only if we are taking a slice of width 0 at the start of an FpVector. This should be a very rare edge case. Dealing with the underflow properly would probably require using saturating_sub or something of that nature, and that has a nontrivial (10%) performance hit.

pub(super) fn min_limb_mask(&self) -> u64

pub(super) fn max_limb_mask(&self) -> u64

pub(super) fn limb_masks(&self) -> (u64, u64)