Struct fp::vector::inner::FqVector

pub struct FqVector<F: Field> {
    pub(super) fq: F,
    pub(super) len: usize,
    pub(super) limbs: Vec<u64>,
}

An FqVector is a vector over a finite field.

Interally, it packs entries of the vectors into limbs. However, this is an abstraction that must not leave the fp library.

Fields
fq: Flen: usizelimbs: Vec<u64>
Implementations

impl<F: Field> FqVector<F>

pub(super) fn scale_helper(&mut self, c: F::ElementContainer)

pub(super) fn entry_helper(&self, index: usize) -> F::ElementContainer

pub(super) fn set_entry_helper( &mut self, index: usize, value: F::ElementContainer, )

pub(super) fn add_helper(&mut self, other: &Self, c: F::ElementContainer)

pub(super) fn add_offset_helper( &mut self, other: &Self, c: F::ElementContainer, offset: usize, )

pub(super) fn add_basis_element_helper( &mut self, index: usize, value: F::ElementContainer, )

pub(super) fn copy_from_slice_helper(&mut self, other: &[F::ElementContainer])

pub(super) fn add_truncate_helper( &mut self, other: &Self, c: F::ElementContainer, ) -> Option<()>

pub(super) fn add_carry_helper<T>( &mut self, other: &Self, c: F::ElementContainer, rest: &mut [T], ) -> bool
where for<'a> &'a mut T: TryInto<&'a mut Self>,

pub(super) fn first_nonzero_helper( &self, ) -> Option<(usize, F::ElementContainer)>

impl<F: Field> FqVector<F>

pub fn new(fq: F, len: usize) -> Self

pub fn from_raw_parts(fq: F, len: usize, limbs: Vec<u64>) -> Self

pub fn new_with_capacity(fq: F, len: usize, capacity: usize) -> Self

pub fn from_slice(fq: F, slice: &[FieldElement<F>]) -> Self

pub fn fq(&self) -> F

pub const fn len(&self) -> usize

pub const fn is_empty(&self) -> bool

pub fn prime(&self) -> ValidPrime

pub fn slice(&self, start: usize, end: usize) -> SliceP<'_, F>

pub fn slice_mut(&mut self, start: usize, end: usize) -> SliceMutP<'_, F>

pub fn as_slice(&self) -> SliceP<'_, F>

pub fn as_slice_mut(&mut self) -> SliceMutP<'_, F>

pub fn add_basis_element(&mut self, index: usize, value: FieldElement<F>)

pub fn entry(&self, index: usize) -> FieldElement<F>

pub fn set_entry(&mut self, index: usize, value: FieldElement<F>)

pub fn iter(&self) -> FqVectorIteratorP<'_, F>

pub fn iter_nonzero(&self) -> FqVectorNonZeroIteratorP<'_, F>

pub fn set_to_zero(&mut self)

pub fn scale(&mut self, c: FieldElement<F>)

pub fn add_offset(&mut self, other: &Self, c: FieldElement<F>, offset: usize)

Add other to self on the assumption that the first offset entries of other are empty.

pub fn add(&mut self, other: &Self, c: FieldElement<F>)

pub fn assign(&mut self, other: &Self)

pub fn assign_partial(&mut self, other: &Self)

A version of FqVector::assign that allows other to be shorter than self.

pub fn is_zero(&self) -> bool

pub(crate) fn limbs(&self) -> &[u64]

pub(crate) fn limbs_mut(&mut self) -> &mut [u64]

pub fn extend_len(&mut self, len: usize)

This function ensures the length of the vector is at least len. See also set_scratch_vector_size.

pub fn set_scratch_vector_size(&mut self, len: usize)

This clears the vector and sets the length to len. This is useful for reusing allocations of temporary vectors.

pub fn copy_from_slice(&mut self, slice: &[FieldElement<F>])

This replaces the contents of the vector with the contents of the slice. The two must have the same length.

pub(crate) fn trim_start(&mut self, n: usize)

Permanently remove the first n elements in the vector. n must be a multiple of the number of entries per limb

pub fn sign_rule(&self, other: &Self) -> bool

pub fn add_truncate(&mut self, other: &Self, c: FieldElement<F>) -> Option<()>

fn add_carry_limb<T>( &mut self, idx: usize, source: u64, c: FieldElement<F>, rest: &mut [T], ) -> bool
where for<'a> &'a mut T: TryInto<&'a mut Self>,

pub fn add_carry<T>( &mut self, other: &Self, c: FieldElement<F>, rest: &mut [T], ) -> bool
where for<'a> &'a mut T: TryInto<&'a mut Self>,

pub fn first_nonzero(&self) -> Option<(usize, FieldElement<F>)>

Find the index and value of the first non-zero entry of the vector. None if the vector is zero.

pub fn density(&self) -> f32