Module fp::vector_inner

Structs

AddShiftLeftData 🔒

AddShiftRightData 🔒

FpVectorIterator

FpVectorNonZeroIteratorP

Iterator over non-zero entries of an FpVector. This is monomorphized over P for significant performance gains.

FpVectorP

An FpVectorP is a vector over $\mathbb{F}_p$ for a fixed prime, implemented using const generics. Due to limitations with const generics, we cannot constrain P to actually be a prime, so we allow it to be any u32. However, most functions will panic if P is not a prime.

SliceMutP

A SliceMutP is a mutable slice of an FpVectorP. This mutably borrows the vector. Since it is a mutable borrow, it cannot implement Copy. However, it has a SliceMutP::copy function that imitates the reborrowing, that mutably borrows SliceMutP and returns a SliceMutP with a shorter lifetime.

SliceP

A SliceP is a slice of an FpVectorP. This immutably borrows the vector and implements Copy