Struct fp::vector_inner::FpVectorP

pub struct FpVectorP<const P: u32> {
    len: usize,
    limbs: Vec<u64>,
}

An FpVectorP is a vector over $\mathbb{F}_p$ for a fixed prime, implemented using const generics. Due to limitations with const generics, we cannot constrain P to actually be a prime, so we allow it to be any u32. However, most functions will panic if P is not a prime.

Interally, it packs entries of the vectors into limbs. However, this is an abstraction that must not leave the fp library.

Fields
len: usizelimbs: Vec<u64>
Implementations

impl FpVectorP<2>

pub fn new(_p: ValidPrime, len: usize) -> FpVector

pub fn new_with_capacity( _p: ValidPrime, len: usize, capacity: usize ) -> FpVector

pub fn from_slice(_p: ValidPrime, slice: &[u32]) -> Self

pub fn num_limbs(_p: ValidPrime, len: usize) -> usize

pub(crate) fn padded_len(p: ValidPrime, len: usize) -> usize

pub fn update_from_bytes(&mut self, data: &mut impl Read) -> Result<()>

pub fn from_bytes( p: ValidPrime, len: usize, data: &mut impl Read ) -> Result<Self>

pub fn to_bytes(&self, buffer: &mut impl Write) -> Result<()>

impl<const P: u32> FpVectorP<P>

pub fn new_(len: usize) -> Self

pub fn from_raw_parts(len: usize, limbs: Vec<u64>) -> Self

pub fn new_with_capacity_(len: usize, capacity: usize) -> Self

pub const fn len(&self) -> usize

pub const fn is_empty(&self) -> bool

pub const fn prime(&self) -> ValidPrime

pub fn slice(&self, start: usize, end: usize) -> SliceP<'_, P>

pub fn slice_mut(&mut self, start: usize, end: usize) -> SliceMutP<'_, P>

pub fn as_slice(&self) -> SliceP<'_, P>

pub fn as_slice_mut(&mut self) -> SliceMutP<'_, P>

pub fn add_basis_element(&mut self, index: usize, value: u32)

pub fn entry(&self, index: usize) -> u32

pub fn set_entry(&mut self, index: usize, value: u32)

pub fn iter(&self) -> FpVectorIterator<'_>

pub fn iter_nonzero(&self) -> FpVectorNonZeroIteratorP<'_, P>

pub fn set_to_zero(&mut self)

pub fn scale(&mut self, c: u32)

pub fn add_offset(&mut self, other: &FpVectorP<P>, c: u32, offset: usize)

Add other to self on the assumption that the first offset entries of other are empty.

pub fn add_offset_nosimd(&mut self, other: &FpVectorP<P>, c: u32, offset: usize)

Add other to self on the assumption that the first offset entries of other are empty.

pub fn add(&mut self, other: &FpVectorP<P>, c: u32)

pub fn add_nosimd(&mut self, other: &FpVectorP<P>, c: u32)

pub fn assign(&mut self, other: &Self)

pub fn assign_partial(&mut self, other: &Self)

A version of FpVectorP::assign that allows other to be shorter than self.

pub fn is_zero(&self) -> bool

pub(crate) fn limbs(&self) -> &[u64]

pub(crate) fn limbs_mut(&mut self) -> &mut [u64]

pub fn extend_len(&mut self, len: usize)

This function ensures the length of the vector is at least len. See also set_scratch_vector_size.

pub fn set_scratch_vector_size(&mut self, len: usize)

This clears the vector and sets the length to len. This is useful for reusing allocations of temporary vectors.

pub fn copy_from_slice(&mut self, slice: &[u32])

This replaces the contents of the vector with the contents of the slice. The two must have the same length.

pub(crate) fn trim_start(&mut self, n: usize)

Permanently remove the first n elements in the vector. n must be a multiple of the number of entries per limb

pub fn sign_rule(&self, other: &Self) -> bool

pub fn add_truncate(&mut self, other: &Self, c: u32) -> Option<()>

fn add_carry_limb<T>( &mut self, idx: usize, source: u64, c: u32, rest: &mut [T] ) -> boolwhere for<'a> &'a mut T: TryInto<&'a mut Self>,

pub fn add_carry<T>(&mut self, other: &Self, c: u32, rest: &mut [T]) -> boolwhere for<'a> &'a mut T: TryInto<&'a mut Self>,

pub fn first_nonzero(&self) -> Option<(usize, u32)>

Find the index and value of the first non-zero entry of the vector. None if the vector is zero.

pub fn density(&self) -> f32